Move execution of commands into their own class
