+++ src/main/java/seedu/coinflip/utils/command/ExitCommand.java
@@ -1,7 +1,5 @@
 package seedu.coinflip.utils.command;
 
-import seedu.coinflip.Coinflip;
-import seedu.coinflip.utils.exceptions.CoinflipFileException;
 import seedu.coinflip.utils.logger.CoinflipLogger;
 import seedu.coinflip.utils.printer.Printer;
 
@@ -9,10 +7,8 @@ import seedu.coinflip.utils.printer.Printer;
  * Handles the command to exit the program.
  */
 public class ExitCommand extends Command {
-    private final Coinflip coinflip;
 
-    public ExitCommand(Coinflip coinflip) {
-        this.coinflip = coinflip;
+    public ExitCommand() {
         CoinflipLogger.fine("ExitCommand created");
     }
 
@@ -20,14 +16,9 @@ public class ExitCommand extends Command {
     // @@author HTY2003
     public void execute() {
         CoinflipLogger.info("Executing exit command");
+
         Printer.printBye();
-        try {
-            coinflip.saveData();
-            CoinflipLogger.info("Game state saved successfully");
-        } catch (CoinflipFileException e) {
-            CoinflipLogger.severe("Failed to save game state: " + e.message);
-            Printer.printException(e);
-        }
+
         CoinflipLogger.info("Exit command execution completed");
     }
 }
