+++ src/main/java/seedu/coinflip/utils/command/FlipCommand.java
@@ -1,25 +1,102 @@
 package seedu.coinflip.utils.command;
 
 import seedu.coinflip.utils.exceptions.CoinflipException;
-import seedu.coinflip.Coinflip;
 import seedu.coinflip.utils.logger.CoinflipLogger;
+import seedu.coinflip.utils.printer.Printer;
+import seedu.coinflip.utils.storage.Storage;
+import seedu.coinflip.utils.userdata.UserData;
+
+import java.util.Random;
 
 /**
  * Handles the command to flip a coin, given a certain bet amount.
  */
 public class FlipCommand extends Command {
     private final String[] words;
-    private final Coinflip coinflip;
+    private final Storage storage;
+    private final UserData userData;
 
-    public FlipCommand(String[] words, Coinflip coinflip) {
+    public FlipCommand(String[] words, UserData userData, Storage storage) {
         this.words = words;
-        this.coinflip = coinflip;
+        this.userData = userData;
+        this.storage = storage;
         CoinflipLogger.fine("FlipCommand created");
     }
 
     @Override
     public void execute() throws CoinflipException {
         CoinflipLogger.info("Executing flip command");
-        coinflip.bet(words);
+        this.bet(words);
+    }
+
+    //@@author wongyihao0506
+    public void bet(String[] words) throws CoinflipException {
+        CoinflipLogger.info("User attempting to flip a coin");
+
+        if (words.length != 2) {
+            CoinflipLogger.warning("Invalid flip command format");
+            throw new CoinflipException(CoinflipException.FLIP_INVALID_FORMAT);
+        }
+
+        if (!words[1].equals("heads") && !words[1].equals("tails")) {
+            CoinflipLogger.warning("Invalid flip choice: " + words[1]);
+            throw new CoinflipException(CoinflipException.FLIP_INVALID_FORMAT);
+        }
+
+        if (userData.betAmount > userData.balance) {
+            CoinflipLogger.warning("Bet amount exceeds balance: bet = "
+                    + userData.betAmount
+                    + ", balance = "
+                    + userData.balance);
+            Printer.printNotEnoughCoins();
+            return;
+        }
+
+        Random random = new Random();
+        String coinFlip = random.nextBoolean() ? "Heads" : "Tails";
+        Boolean outcome = coinFlip.equalsIgnoreCase(words[1]);
+
+        if (outcome) {
+            userData.balance += userData.betAmount;
+            increaseWinCount();
+            increaseTotalWinnings(userData.betAmount);
+            CoinflipLogger.info("User won "
+                    + userData.betAmount
+                    + " coins. New balance: "
+                    + userData.balance);
+        } else {
+            userData.balance -= userData.betAmount;
+            increaseLoseCount();
+            increaseTotalLosses(userData.betAmount);
+            CoinflipLogger.info("User lost " + userData.betAmount + " coins. New balance: " + userData.balance);
+        }
+
+        Printer.printBetAmount(userData.betAmount);
+        Printer.printFlipOutcome(coinFlip, outcome, userData.betAmount);
+        Printer.printBalance(userData.balance);
+
+        assert userData.balance >= 0 : "balance should be more than or equal to 0";
+
+        storage.saveData(userData);
+    }
+
+    //@@author CRL006
+    private void increaseWinCount() {
+        userData.winCount += 1;
+    }
+
+    //@@author CRL006
+    private void increaseLoseCount() {
+        userData.loseCount += 1;
+    }
+
+    //@@author CRL006
+    private void increaseTotalWinnings(int winnings) {
+        userData.totalWinnings += winnings;
+    }
+
+    //@@author CRL006
+    private void increaseTotalLosses(int losses) {
+        userData.totalLosses += losses;
     }
 }
