+++ src/main/java/seedu/coinflip/utils/command/HelpCommand.java
@@ -8,7 +8,8 @@ import seedu.coinflip.utils.printer.Printer;
  */
 public class HelpCommand extends Command {
 
-    public HelpCommand() {}
+    public HelpCommand() {
+    }
 
     public void execute() {
         CoinflipLogger.info("Executing help command");
